requirements.txt
torch
torchvision
transformer_lens
wandb
tqdm
pandas
captcha
Pillow
pytesseract
numpy
einops
jaxtyping